Revert "Rwa 2310 spring boot upgrade (#745)"
This reverts commit 7653f0ee0b68297ed063453b65df3de728fb264b.
