+++ src/functionalTest/java/uk/gov/hmcts/reform/wacaseeventhandler/config/CcdEventPublisherConfiguration.java
@@ -2,11 +2,12 @@ package uk.gov.hmcts.reform.wacaseeventhandler.config;
 
 import com.azure.messaging.servicebus.ServiceBusClientBuilder;
 import com.azure.messaging.servicebus.ServiceBusSenderClient;
-import jakarta.annotation.PreDestroy;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import javax.annotation.PreDestroy;
 
 @Configuration
 @ConditionalOnExpression("${azure.servicebus.enableASB-DLQ:true}")
