+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/config/OpenAPIConfiguration.java
@@ -1,13 +1,10 @@
 package uk.gov.hmcts.reform.wacaseeventhandler.config;
 
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.models.ExternalDocumentation;
-import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.Operation;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.info.License;
 import io.swagger.v3.oas.models.media.StringSchema;
 import io.swagger.v3.oas.models.parameters.Parameter;
+import org.springdoc.core.GroupedOpenApi;
 import org.springdoc.core.customizers.OperationCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -17,14 +14,12 @@ import org.springframework.web.method.HandlerMethod;
 public class OpenAPIConfiguration {
 
     @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-            .info(new Info().title("wa-case-event-handler Service")
-                      .version("v1.0.0")
-                      .license(new License().name("MIT").url("https://opensource.org/licenses/MIT")))
-            .externalDocs(new ExternalDocumentation()
-                              .description("README")
-                              .url("https://github.com/hmcts/wa-case-event-handler"));
+    public GroupedOpenApi publicApi(OperationCustomizer customGlobalHeaders) {
+        return GroupedOpenApi.builder()
+            .group("case-event-handler-public")
+            .pathsToMatch("/**")
+            .addOperationCustomizer(customGlobalHeaders)
+            .build();
     }
 
     @Bean
