+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/config/RestTemplateConfiguration.java
@@ -1,15 +1,15 @@
 package uk.gov.hmcts.reform.wacaseeventhandler.config;
 
-import org.apache.hc.client5.http.classic.HttpClient;
-import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.web.client.RestTemplate;
-import org.zalando.logbook.httpclient5.LogbookHttpRequestInterceptor;
-import org.zalando.logbook.httpclient5.LogbookHttpResponseInterceptor;
+import org.zalando.logbook.httpclient.LogbookHttpRequestInterceptor;
+import org.zalando.logbook.httpclient.LogbookHttpResponseInterceptor;
 
 import java.util.function.Supplier;
 
@@ -33,9 +33,9 @@ public class RestTemplateConfiguration {
     class RequestFactorySupplier implements Supplier<ClientHttpRequestFactory> {
         @Override
         public ClientHttpRequestFactory get() {
-            HttpClient client = HttpClientBuilder.create()
-                .addRequestInterceptorFirst(logbookHttpRequestInterceptor)
-                .addResponseInterceptorFirst(logbookHttpResponseInterceptor)
+            CloseableHttpClient client = HttpClientBuilder.create()
+                .addInterceptorFirst(logbookHttpRequestInterceptor)
+                .addInterceptorFirst(logbookHttpResponseInterceptor)
                 .build();
             return new HttpComponentsClientHttpRequestFactory(client);
         }
