+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/config/SecurityConfiguration.java
@@ -2,12 +2,11 @@ package uk.gov.hmcts.reform.wacaseeventhandler.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.web.authentication.preauth.AbstractPreAuthenticatedProcessingFilter;
 import uk.gov.hmcts.reform.authorisation.filters.ServiceAuthFilter;
 
@@ -20,7 +19,7 @@ import static org.springframework.security.config.http.SessionCreationPolicy.STA
 @Configuration
 @ConfigurationProperties(prefix = "security")
 @EnableWebSecurity
-public class SecurityConfiguration {
+public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 
     private final List<String> anonymousPaths = new ArrayList<>();
     @Autowired
@@ -30,15 +29,15 @@ public class SecurityConfiguration {
         return anonymousPaths;
     }
 
-    @Bean
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        return (web) -> web.ignoring().requestMatchers(anonymousPaths.toArray(String[]::new));
+    @Override
+    public void configure(WebSecurity web) {
+        web.ignoring().mvcMatchers(
+            anonymousPaths.toArray(String[]::new)
+        );
     }
 
-    @Bean
-    @SuppressWarnings("PMD.SignatureDeclareThrowsException")
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
         http
             .addFilterBefore(serviceAuthFilter, AbstractPreAuthenticatedProcessingFilter.class)
             .sessionManagement().sessionCreationPolicy(STATELESS)
@@ -47,7 +46,5 @@ public class SecurityConfiguration {
             .formLogin().disable()
             .logout().disable()
             .csrf().disable();
-
-        return http.build();
     }
 }
