+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/config/executors/CcdMessageProcessorExecutor.java
@@ -1,7 +1,5 @@
 package uk.gov.hmcts.reform.wacaseeventhandler.config.executors;
 
-import jakarta.annotation.PostConstruct;
-import jakarta.annotation.PreDestroy;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -11,6 +9,9 @@ import uk.gov.hmcts.reform.wacaseeventhandler.clients.DatabaseMessageConsumer;
 
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+
 
 @Component
 @Slf4j
